+++ workspaces/copilot/plugins/copilot-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-copilot-common
 
+## 0.16.0
+
+### Minor Changes
+
+- d2133ca: Backstage version bump to v1.46.1
+
 ## 0.15.0
 
 ### Minor Changes