+++ workspaces/entity-feedback/plugins/entity-feedback-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage-community/plugin-entity-feedback-backend
 
+## 0.16.0
+
+### Minor Changes
+
+- a3c15ac: Backstage version bump to v1.46.1
+
+### Patch Changes
+
+- Updated dependencies [a3c15ac]
+  - @backstage-community/plugin-entity-feedback-common@0.12.0
+
 ## 0.15.0
 
 ### Minor Changes