+++ workspaces/explore/plugins/explore/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage-community/plugin-explore
 
+## 0.16.0
+
+### Minor Changes
+
+- 9db38a1: Backstage version bump to v1.46.1
+
+### Patch Changes
+
+- Updated dependencies [9db38a1]
+  - @backstage-community/plugin-explore-common@0.11.0
+  - @backstage-community/plugin-explore-react@0.11.0
+
 ## 0.15.0
 
 ### Minor Changes