+++ workspaces/github-pull-requests-board/plugins/github-pull-requests-board/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-github-pull-requests-board
 
+## 0.13.0
+
+### Minor Changes
+
+- 49c61e1: Backstage version bump to v1.46.1
+
 ## 0.12.0
 
 ### Minor Changes