+++ workspaces/lighthouse/plugins/lighthouse-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-lighthouse-common
 
+## 0.14.0
+
+### Minor Changes
+
+- 5ea7ce0: Backstage version bump to v1.46.1
+
 ## 0.13.0
 
 ### Minor Changes