+++ workspaces/linguist/plugins/linguist-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-linguist-common
 
+## 0.15.0
+
+### Minor Changes
+
+- 7209097: Backstage version bump to v1.46.1
+
 ## 0.14.0
 
 ### Minor Changes