+++ workspaces/sonarqube/packages/backend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # backend
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [7a2bd54]
+  - @backstage-community/plugin-sonarqube-backend@0.16.0
+  - app@0.0.31
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [5125502]
+  - @backstage-community/plugin-sonarqube-backend@0.15.0
+  - app@0.0.30
+
 ## 0.0.18
 
 ### Patch Changes