+++ workspaces/sonarqube/plugins/sonarqube-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage-community/plugin-sonarqube-backend
 
+## 0.16.0
+
+### Minor Changes
+
+- 7a2bd54: Backstage version bump to v1.46.1
+
+## 0.15.0
+
+### Minor Changes
+
+- 5125502: Backstage version bump to v1.45.1
+
 ## 0.14.0
 
 ### Minor Changes