+++ workspaces/todo/plugins/todo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-todo
 
+## 0.15.0
+
+### Minor Changes
+
+- a364574: Backstage version bump to v1.46.1
+
 ## 0.14.0
 
 ### Minor Changes